Make the moon orbit the earth
The earth was spinning to fast to see so I changed it's angle counter
from int to float64 and step it by 0.02 each time.  It turns out that
when looking from above, the earth and moon rotate anti-clockwise, so I
set the rotation to decrement.

I expected rotation to be done with trigonometry but instead it's done
with matrix movements, move the image so that the your desired point is
at the screen's origin, apply rotation and then position your image on
the screen.  By re-using the earth's rotation logic it was easy to make
the moon rotate so that it always has the same side facing the earth,
i.e. tidal locking.

The moon doesn't orbit at the same speed the earth rotates, obviously, a
rotation is one day and an orbit is more like a month (~27 days) but a
27x speed difference looked really bad in the game, so I rounded up and
dropped a digit to make it 3.  It looks pretty cool now.

In real life the moon is roughly 10 moons' distance from the earth but
that would look bad and wouldn't fit in the screen, so I positioned it 2
moons' away.  It feels a little bit close but it fits on the screen.