Move earth and asteroid update logic to their own functions
They each have their own independent GeoM instead of reusing one and
resetting it every time, it's a little bit more code but the main Update
function is now very simple and just calls out to the other objects to
update, it could probably be made into a loop.  The main thing I don't
like is the function signature, that some of them receive earth (which
they need for some distance calculations) and some of others not.