Fix Radius calculation
The Radius was set to the width of the image, which is its diameter, and
was then halved everywhere where objects needed to be centred.  I've
done that halving at initialisation now, so it's really the radius and you
don't need to divide it by two anywhere else in the code any more.