Add debug function to visually debug with a ruler on the screen
Some of the maths got hard, so this debugging function draws some
squares and lines on the screen to help me measure where things are when
they aren't where they should be, to help me find where I made mistakes
in the code.  It also prints out the mouse X, Y and straight line
distance from the centre of the screen.