Fix collision calculation between earth and asteroid
It seems I made a maths error early on in calculating the starting
distance for the asteroid but didn't realise it and then compensated for
it with guesswork at later stages like drawing and collision
calculation.  The assumption at the time was probably that distance
represents the distance from the edge of the earth but this made
calculations elsewhere harder.

Now instead of the asteroid starting 2 moons distance away plus some
magic distance addition happening elsewhere, it's positioned 2 earth
distances away.  The collision calculation was simplified so that it
evaluates to true when it reaches zero, not some other calculated value.
Also the drawn position of the asteroid is adjusted so that at distance
zero half of its body has overlapped the earth so that it _feels_ more
like a collision.  For the last part I added a Radius field to the
asteroid too, so that I could calculate how much to readjust it.