Hide mouse cursor in the game
In future it would be cool to add some build-time conditions so that
when building for desktop (where it's supported) it'll lock the cursor
to the game and for other platforms like web (where you can't lock but
at least hide) or Android where there is no cursor it sets it a bit
differently.  That can come later though.