Make an exploding animation when the asteroid impacts the Earth
It took a while to draw all the frames for this animation, there are 7
so far and I might tweak them later.  Currently in a push to just get it
working all the state and everything about the explosion is stored in
extra globals with some long drawing logic for finding the right frame
of animation too.

I'm imagining the explosion state information would probably belong to
the asteroids (especially since in future each asteroid may have its own
explosion) and the frame logic could be moved to a method of the
explosion object itself.