Centre explosion on Asteroid impact site
This was done by copy-pasting across the angle calculation logic from
the Asteroid Update function, then it turned out the distance
calculation is much simpler because it doesn't move, so it turned out to
be a fairly small code change.  Could definitely still be improved
though, e.g. only calculating this if explosion is in progress, not for
every tick.