Calculate collision between crosshair and asteroid
The collision detection is used by comparing the distance between the
centres to their combined radii[1].  This function could do with some
tests.  Because this produced different coordinates than where the
asteroid image was appearing I dropped the old positioning logic and now
position it based on this x, y.  Next step is to make the code actually
do something with this new knowledge of collisions.

[1]: http://jeffreythompson.org/collision-detection/circle-circle.php