Start off with 3 asteroids instead of just one
This required a lot of changes, now instead of handling each object
one-by-one copy-paste style, the asteroids are all in a slice, with
their own Asteroids type.  Luckily because it's a type like any other, I
can give it Update & Draw methods, so it can still be called up by the
main Update & Draw functions without any modification to them.
Internally the Asteroids' methods just loop over its slice, calling the
same method on each one, but that's abstracted away neatly from the
outside.

The one place that this isn't abstracted away still is handling clicks
when the player shoots, I'll probably move that to the Asteroid update
function.  This has required shuffling about a lot of states that were
in other places before, and sometimes duplicating things to make them
accessible when you don't have access by name to a thing that used to be
a variable but now it's not.