Share asteroid & explosion image across object instances
I noticed that when starting the game with many asteroids it loads their
images many times but once would be enough.  I wrote a new constructor
that takes an image instead of a string so that you can pass it an image
loaded in advance to avoid loading it more than once.