Add game restart logic with fresh asteroids
If the player clicks the mouse during game over state, it generates a
new field of asteroids and resets the game states to their default
values, including the Game Over state, thus enable the game to carry on
with new asteroids and zero score.

The NewAsteroids constructor is the same code that was in main, just
moved to its own function.