Double the asteroids when the previous wave has been defeated
If the game is not over (earth destroyed) and all the asteroids are dead
then the wave has been defeated and we need to ramp up the difficulty.
For lack of a better idea I'm just doubling the number of asteroids.
Their distance, speed etc doesn't very though, so it means a lot of them
arrive at roughly the same time, making it quite challenging.