Start asteroids off screen
I picked this 3.5 value through experimentation as just big enough to
have them starting off the edge of the screen.  With them starting on
screen the start of a new wave would be jarring, with a sudden flash of
asteroids, now instead they start leaking in ominously like a real wave.