Set up a first unit test
This test doesn't do anything useful yet except to show that test
running works fine and the CI pipeline can handle it correctly.
Consider it a dummy that needs to be replaced with some useful test
cases.  I picked overlaps function because it's used for collision
detection for all the objects, so it would be pretty criritcal to verify
that it's working right.