Count asteroids down instead of up
Instead of counting each hit asteroid, show how many there are and
decrement it for each one you hit, this also includes hitting them with
the moon.  Counting the moon too introduced an interesting bug that the
moon can make it explode if it hits it while it's exploding because it
only checks if it's Alive before triggering explosion, not checking if
it's already explosion.  That's fixed now though.