Draw a laser sight from the moon to the crosshair
This does nothing, it's just visuals but it helps include the moon into
the game a bit more and justify why there's a giant red crosshair in the
sky.  One unintentional advantage is that it helps track the position of
the moon in relation to you, so that you can coordinate.