Override main game "constants" from an ini file
This is mostly so that beta testers can experiment with different values
to find what provides a nice game experience, I don't intent this for
end users of the production version.  I've used the go-ini library for
parsing the ini files, it has many more features than I need at the
moment so I might consider using something simpler in future but I don't
have time to spend researching libraries at the moment and it works fine
for me.  I'll clean up the config usage code later.

I've provided an example file with the default values at the time of
writing.  The name of the config file the game looks for is hard-coded
as: lunar-defence.ini.