Add a cooldown time as a penalty for missed shots
If the player clicks to shoot but doesn't hit an asteroid, that counts
as a miss.  A new timer is started in a goroutine that currently waits
one second before setting the crosshair to cooled-down.  During that
time you can still click but it won't shoot.

Would be cool if this had an animation, e.g. an explosion on the moon to
show that you've damaged yourself with your bad aiming.  Another idea is
to freeze the crosshair to emphasise that it's not usable.