Disallow shooting during the breathing before new waves
There's nothing to shoot anyway, but the issue this fixes is that if you
shoot during the "take a breath" stage then the text about cooling down
and the text about the new wave coming will both be shown at the same
time, drawn over each other.  This is fixed by not letting the player
shoot during the breathing time.