Animate an explosion on the moon when you miss a shot
The existing explosion object was reused from the asteroids.  It has an
asteroid parameter but only the centre coordinates were used, so I
changed it so that only the coordinates are passed, and then passed it
the moon's coordinates when exploding on a miss.  The shooting/missing
logic was also updated so that it shows the