Don't count clicking for game restart as a miss
Technically you are not clicking on an asteroid, but also there are no
asteroids to click on when the game is over, so don't count this click
as a miss.  For this to work it required moving the Game Over handling
to after all the other sprites have been handled, otherwise you're on
the game over screen, set the value to true and then in the crosshair
update function that same click is also handled.