Move game entities set up to its own function
This is cluttering the main function with what is just lots of
repetitive image loading and entity set-up.  It'd be easier to read the
main function as an overview if this was moved to its own function.

Having NewGame set up the objects in its own function is also a step
towards doing this loading asynchronously so that on slower machines
that main loop can start sooner and show a loading screen until all
entities are set up.