Implement loading screen and start screen
Initialising the games entities is now done in a separate go routine
(hopefully concurrently) to allow the main loop to start sooner and show
the player the loading text instead of waiting on a blank screen.

Once everything is loaded, instead of starting the game immediately
I've made a start screen which is basically the Earth and Moon spinning
but no asteroids yet, i.e. wave zero.  In this state clicking the screen
will start wave 1.

The only thing from game initialisation that's needed to display the
loading screen is the font face, so that's been moved out of the
NewGame() function in to a loadFont() function so that I can at least
show that text until everything else is loaded.