Add a turret on the moon to show what's shooting
A new turret image has been added as a sub-entity of the moon.  It's the
weapon that's firing the lasers and it always rotates to face towards
the crosshair so that it's aligned with the laser beam that shoots out
when the player clicks.

Today I learned about the math.Atan2() function which calculates an
Angle given a coordinate.  The input coordinate I provide it is the
Crosshair's coordinates subtracted from the Turret's own coordinates.