Add infinite looping game music
As I was searching for eerie space themes music I came across electronic
music producer Nihilore.  The track, the water and the well, seemed like
a really good fit for this game, atmosphere-wise, it's exactly what I
had in mind, so I've included it with credits on the opening screen.

This is the only sound file in the game that loops continuously, so that
needed a bit of extra work.  The documentation for the infite loop was
not entirely clear, first I thought you need to put the player in it.
After a while of comparing types I realised that you can wrap your file
stream in this infinite loop before passing it to the player to make it
infinite, it wasn't obvious, but it's very convenient and elegant.

The only downside is even at reduced quality the audio file is huge and
the bundle size has more than doubled.  I goess I'll make a separate
release for web with no music.