Only start music when the first wave starts
If the starts immidiately when the game opens then it's a pretty
annoying experience on web, because I set the page to launch the game on
load, so the music would just start playing immidiately (except in
Chrome where you only get an audio context after the first click).  To
solve this, I moved the NewSounds() initialisation to inside the main
Update() function where it has the logic for starting the first wave,
when the player clicks to start the game for the first time.

This solves the problem perfectly and also delays loading the music
files until that time.  The pro is that you see the loading screen for
less time before it shows you the spinning Earth screen but the con is
that there is a slight jarring delay as you click for the first time
because it's loading the audio files, including the fairly large music
sound track.