Use xvfb for fake X11 in Travis
When the tests try to run in Travis it immidiatly crashes because
there's no DISPLAY variable set for the game to start up in, that's
because there really is no display.  According to the Travis GUI Testing
doc https://docs.travis-ci.com/user/gui-and-headless-browsers/ you can
use xvfb instead.  It's the X Virtual Framebuffer, which lets your
program draw offscreen to immitate having a display.

Probably I could've also solved this by splitting my game into 2
separate packages:
- main: which sets up the game and starts it up
- objects: which contains the logic for the game objects and their
  interactions and is what really needs to be tested
And because the objects tests don't require the game to start they would
run fine.  At some point it would probably be good to test that the game
starts fine though, e.g. it might compile but crash if there are files
missing from the statikFS because it only loads those files at runtime.

So, I've turned on xvfb for this repo, that means adding xvfb to the
list  of services and putting xvfb-run before the test command, because
that's the one that requires the display.