Add window icon to the game
This makes so that when the game is running it has an app icon.  I
copied in the high-resolution moon-icon.png that's used for the MacOS
app and set that as the icon.  This made the game crash with Ebiten
v2.0.0 so I updated it to v2.2.4 and it works fine.